+++ .github/workflows/update-packages-list.yml
@@ -10,7 +10,7 @@ concurrency:
   cancel-in-progress: true
 
 jobs:
-  updage-readme:
+  update-readme:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
@@ -19,7 +19,7 @@ jobs:
           node-version: 16
           cache: 'npm'
       - run: npm ci
-      - run: node build-package-list.js
+      - run: node update-readme.js
       - run: git config user.name github-actions
       - run: git config user.email github-actions@github.com
       - run: git add -A
